requirements.txt
aiogram>=3.7.0
httpx>=0.27.2
pydantic>=2.9.2
python-dotenv>=1.0.1
vosk>=0.3.45
imageio-ffmpeg>=0.5.1
